+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 Extends support of the SDK to OneView REST API version 800, 1000 and 1200.
 
 #### Features supported
-- Hypervisor Managers
 - Certificates Server
+- Hypervisor Managers
 
 # 5.0.0
 #### Notes