+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 5.1.0 (unreleased version)
+# 5.1.0
 #### Notes
 Extends support of the SDK to OneView REST API version 800, 1000 and 1200.